+++ pygeos/test/test_predicates.py
@@ -90,3 +90,16 @@ def test_equals_exact_tolerance():
     # default value for tolerance
     assert pygeos.equals_exact(p1, p1).item() is True
     assert pygeos.equals_exact(p1, p2).item() is False
+
+
+def test_relate():
+    p1 = pygeos.points(0, 0)
+    p2 = pygeos.points(1, 1)
+    actual = pygeos.relate(p1, p2)
+    assert isinstance(actual, str)
+    assert actual == "FF0FFF0F2"
+
+
+@pytest.mark.parametrize("g1, g2", [(point, None), (None, point), (None, None)])
+def test_relate_none(g1, g2):
+    assert pygeos.relate(g1, g2) is None